Move all inspector implementation to datainspectortool.py
Merge datainspectorline.py into datainspectortool.py and do some
further code and docstring cleaning.
Also add a taurus-less example in the main section.
